pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slipp_plus"
version = "0.1.0"
description = "SLiPP++ Day 1: 10-class softmax reformulation of Chou et al. 2024"
authors = [{ name = "Filip Rumenovski" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
dependencies = [
  # Paper-pinned
  "scikit-learn==1.3.1",
  "numpy>=1.26,<2.0",
  # Data + validation
  "polars>=1.0,<2.0",
  "pandas>=2.0,<3.0",
  "pandera>=0.20,<0.25",
  "pydantic>=2.5,<3.0",
  "pyarrow>=15.0",
  "openpyxl>=3.1",
  # Models
  "xgboost>=2.0,<3.0",
  "lightgbm>=4.0,<5.0",
  # Viz
  "matplotlib>=3.8,<4.0",
  "seaborn>=0.13,<1.0",
  # Plumbing
  "pyyaml>=6.0",
  "typer>=0.12",
  "tqdm>=4.0",
  "joblib>=1.3",
  # Day 7+ scratch path deps (light, kept in base so download.py imports work)
  "biopython==1.83",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-xdist>=3.5",
  "ruff>=0.5",
  "mypy>=1.10",
]
scratch = [
  # Only needed if you run the from-scratch reproduction (src/pocket_extraction.py).
  # biobb_vs wraps fpocket; install system fpocket separately.
  "biobb-vs>=4.0",
]

[project.scripts]
slipp_plus = "slipp_plus.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/slipp_plus"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP", "SIM", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "E402"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
packages = ["slipp_plus"]

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
